Movendo AMQPRouteEntryPointImpl para um pacote separado
